// Aplicacion-Pc-Front-End/src/Interfaces/Computadoras/ComputadorasInterfaces.ts
export interface Computadoras {
  _id: string;
  name: string;
  Imagen: string;
  Imagenes: string[];
  Contador: number;
  PrecioUnitario: number;
  Procesador: string;
  MemoriaRam: string;
  Categorias: string;
  Marca: string;
  PlacaDeVideo: string;
  DiscoDuro: string;
  SistemaOperativo: string;
  Descripcion: string;
  Precio: number;
  Stock: number;
  Identificador: string;
  Destacado: boolean;
}