// Aplicacion-Pc-Front-End/src/Interfaces/Perifericos/Perifericos.ts
export interface Perifericos {
  _id: string;
  name: string;
  Imagen: string;
  PrecioUnitario: number;
  Contador: number;
  Imagenes: string[];
  Descripcion: string;
  Marca: string;
  Tipo: string;
  Precio: number;
  Stock: number;
  IdCarrusel: string;
  Identificador: string;
  Destacado: boolean;
  TipoDeConexion: string;
}